Update README.md
# Polynomial_Regression

### Summary of the analysis
In this project, different techniques and methods were explored to predict the the amount of someone pays or would pay for insurance com would issue based on ['age', 'sex', 'bmi', 'children', 'smoker', and 'region'] of an individual. We tried to make sense out of the use of polynomial regression , Ridge and Lasso regressions. Indeed,polynomial regression performed best amongst the three models. With the final result, a general template can be made in same manner with our regression analysis repository using the prediciting features in this case of an issurance company.

### General description
A polynomial regression model is a machine learning model that can capture non-linear relationships between variables by fitting a non-linear regression line, which may not be possible with simple linear regression. It is used when linear regression models may not adequately capture the complexity of the relationship. www.analyticsvidhya.com